+++ src/main/java/com/example/java_group_11_online_store_ayday_mirbekkyzy/Service/ProductsService.java
@@ -30,7 +30,7 @@ public class ProductsService {
         }
     }
 
-    public Page<ProductsDTO> priceSearch(Integer more,Integer less,Pageable pageable){
+    public Page<ProductsDTO> priceSearch(Float more,Float less,Pageable pageable){
         Page<Products> products = productsRepository.findProductsByPriceGreaterThanAndPriceIsLessThan(more,less,pageable);
         return products.map(ProductsDTO::from);
     }
